services/rag/pyproject.toml
[project]
name = "rag-builder"
version = "0.1.0"
description = "One-time script: parse resume PDF → embed → SQLite for Lambda RAG"
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.12.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-readers-file>=0.4.0",
    "pymupdf>=1.25.0",
    "openai>=1.0.0",
    "numpy>=1.26.0",
]

[tool.setuptools]
py-modules = ["build_rag"]
